Clippy and cleanup

+++ src/components/experience_container.rs
@@ -1,3 +1,4 @@
+#![allow(clippy::derive_partial_eq_without_eq)]
 use yew::{html::ChildrenRenderer, prelude::*};
 
 #[derive(Clone, PartialEq, Properties)]
@@ -25,9 +26,9 @@ impl Component for ExperienceContainer {
   }
 
   fn view(&self, ctx: &Context<Self>) -> Html {
-    let title_styles = if ctx.props().title.chars().count() > 28 as usize {
+    let title_styles = if ctx.props().title.chars().count() > 28 {
       classes!("font-mulish text-l md:text-xl".to_string())
-    } else if ctx.props().title.chars().count() > 22 as usize {
+    } else if ctx.props().title.chars().count() > 22 {
       classes!("font-mulish text-2xl".to_string())
     } else {
       classes!("font-mulish text-3xl".to_string())
